remove duplicate function
